Added a list for weighing.

+++ app/src/main/java/com/kudashov/rabbits_farm/adapters/delegates/TaskDelegate.kt
@@ -2,6 +2,6 @@ package com.kudashov.rabbits_farm.adapters.delegates
 
 interface TaskDelegate {
     fun confirmSimpleTask(id: Int)
-    fun confirmSlaughterInspectionTask(id: Int, weights: List<Int>)
+    fun confirmSlaughterInspectionTask(id: Int, weights: List<Double>)
     fun confirmDepositionFromMotherTask(id: Int, countMales: Int)
 }
