+++ app/src/main/java/com/kudashov/rabbits_farm/data/source/TaskProvider.kt
@@ -21,8 +21,8 @@ interface TaskProvider {
     fun confirmSlaughterInspectionTask(
         token: String,
         id: Int,
-        weights: List<Int>)
-            : Observable<BaseResponse>
+        weights: List<Double>
+    ): Observable<BaseResponse>
 
     fun confirmDepositionFromMotherTask(
         token: String,
